Movie-list section into SASS
