+++ src/sass/partials/_movie-list.scss
@@ -2,101 +2,96 @@
 @use '../utils/container';
 @use '../utils/mediaquery';
 
-.movie-list-section {
-  @extend .container;
-  background-color: variables.$background-color-lighter;
-  padding-bottom: 40px;
-  padding-top: 20px;
-  @include mediaquery.tablet {
-    padding: 59px 32px 60px;
+.movie-list {
+  &-section {
+    @extend .container;
+    background-color: variables.$background-color-lighter;
+    padding-bottom: 40px;
+    padding-top: 20px;
+    @include mediaquery.tablet {
+      padding: 59px 32px 60px;
+    }
+    @include mediaquery.desktop {
+      padding: 60px 32px;
+    }
   }
-  @include mediaquery.desktop {
-    padding: 60px 32px;
-  }
-}
-
-.movie-list-container {
-  background-color: variables.$background-color-lighter;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin: 0 auto;
-  width: 100%;
-  max-width: 480px;
-  @include mediaquery.tablet {
-    flex-direction: row;
-    flex-wrap: wrap;
-    gap: 32px;
-    max-width: 768px;
-  }
-  @include mediaquery.desktop {
-    column-gap: 16px;
-    max-width: 1280px;
-    row-gap: 32px;
+  &-container {
+    background-color: variables.$background-color-lighter;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 0 auto;
+    width: 100%;
+    max-width: 480px;
+    @include mediaquery.tablet {
+      flex-direction: row;
+      flex-wrap: wrap;
+      gap: 32px;
+      max-width: 768px;
+    }
+    @include mediaquery.desktop {
+      column-gap: 16px;
+      max-width: 1280px;
+      row-gap: 32px;
+    }
   }
 }
 
 .movie-card {
   border-radius: 5px;
   transition: transform 1000ms ease;
-}
-
-.movie-card:focus,
-.movie-card:hover {
-  transform: scale(1.05);
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.12), 0 4px 4px rgba(0, 0, 0, 0.06),
-    1px 4px 6px rgba(0, 0, 0, 0.16);
-}
-
-.movie-card:not(:last-child) {
-  margin-bottom: 20px;
-  @include mediaquery.tablet {
-    margin-bottom: 0;
+  &:focus,
+  &:hover {
+    transform: scale(1.05);
+    box-shadow: 0 1px 1px rgba(0, 0, 0, 0.12), 0 4px 4px rgba(0, 0, 0, 0.06),
+      1px 4px 6px rgba(0, 0, 0, 0.16);
   }
-}
-
-.movie-card__poster {
-  border-radius: 5px;
-  margin-bottom: 10px;
-  @include mediaquery.tablet {
-    margin-bottom: 8px;
-    width: 336px;
-    height: 455px;
+  &:not(:last-child) {
+    margin-bottom: 20px;
+    @include mediaquery.tablet {
+      margin-bottom: 0;
+    }
   }
-  @include mediaquery.desktop {
-    margin-bottom: 12px;
-    width: 393px;
-    height: 574px;
+  &__poster {
+    border-radius: 5px;
+    margin-bottom: 10px;
+    @include mediaquery.tablet {
+      margin-bottom: 8px;
+      width: 336px;
+      height: 455px;
+    }
+    @include mediaquery.desktop {
+      margin-bottom: 12px;
+      width: 393px;
+      height: 574px;
+    }
   }
-}
-
-.movie-card__info {
-  border-bottom-left-radius: 5px;
-  border-bottom-right-radius: 5px;
-  background-color: variables.$background-color-lighter;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 1.33;
-  max-width: 280px;
-  padding-bottom: 5px;
-  @include mediaquery.tablet {
-    max-width: 336px;
+  &__info {
+    border-bottom-left-radius: 5px;
+    border-bottom-right-radius: 5px;
+    background-color: variables.$background-color-lighter;
+    font-size: 12px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 1.33;
+    max-width: 280px;
+    padding-bottom: 5px;
+    @include mediaquery.tablet {
+      max-width: 336px;
+    }
+    @include mediaquery.desktop {
+      font-size: 20px;
+      line-height: 1.2;
+      max-width: 394px;
+    }
   }
-  @include mediaquery.desktop {
-    font-size: 20px;
-    line-height: 1.2;
-    max-width: 394px;
+  &__title {
+    color: variables.$text-color-darker;
+    text-transform: uppercase;
   }
-}
-
-.movie-card__title {
-  color: variables.$text-color-darker;
-  text-transform: uppercase;
-}
-
-.movie-card__genre {
-  color: variables.$active-color;
+  &__genre {
+    color: variables.$active-color;
+  }
 }
 
 .movie-card--placeholder {
@@ -118,12 +113,11 @@
     width: 394px;
     height: 634px;
   }
-}
-
-.movie-card--placeholder__title {
-  color: variables.$active-color;
-  font-size: 45px;
-  font-weight: 750;
-  line-height: 1.33;
-  text-align: center;
+  &__title {
+    color: variables.$active-color;
+    font-size: 45px;
+    font-weight: 750;
+    line-height: 1.33;
+    text-align: center;
+  }
 }
